Fix cast operation in RgbeInfo

+++ src/main/java/org/apache/commons/imaging/formats/rgbe/RgbeInfo.java
@@ -125,7 +125,7 @@ final class RgbeInfo implements Closeable {
                     if (0 == mantissa) {
                         out[channel][pos] = 0;
                     } else {
-                        final float mult = (float) Math.pow(2, mantissa - (128 + 8));
+                        final float mult = (float) Math.pow(2, (double) mantissa - (128 + 8));
                         out[channel][pos] = ((rgbe[p + channelOffset] & 0xff) + 0.5f) * mult;
                     }
                 }
